chore: extracted engine
