+++ src/engine/inference.ts
